core: Move Mavlink FTP logic into core
This way FTP can be used in the plugin but also by other plugins such as
component information server.
