+++ src/mavsdk/core/system_impl.h
@@ -5,6 +5,7 @@
 #include "mavlink_parameters.h"
 #include "mavlink_command_receiver.h"
 #include "mavlink_command_sender.h"
+#include "mavlink_ftp.h"
 #include "mavlink_message_handler.h"
 #include "mavlink_mission_transfer.h"
 #include "mavlink_request_message_handler.h"
@@ -282,6 +283,8 @@ public:
 
     MAVLinkMissionTransfer& mission_transfer() { return _mission_transfer; };
 
+    MavlinkFtp& mavlink_ftp() { return _mavlink_ftp; };
+
     RequestMessage& request_message() { return _request_message; };
 
     void intercept_incoming_messages(std::function<bool(mavlink_message_t&)> callback);
@@ -426,6 +429,7 @@ private:
 
     MAVLinkMissionTransfer _mission_transfer;
     RequestMessage _request_message;
+    MavlinkFtp _mavlink_ftp;
 
     std::mutex _plugin_impls_mutex{};
     std::vector<PluginImplBase*> _plugin_impls{};
